move internal classes to their own file
